+++ source/source_test.go
@@ -25,92 +25,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestParseTemplate(t *testing.T) {
-	for _, tt := range []struct {
-		name                     string
-		annotationFilter         string
-		fqdnTemplate             string
-		combineFQDNAndAnnotation bool
-		expectError              bool
-	}{
-		{
-			name:         "invalid template",
-			expectError:  true,
-			fqdnTemplate: "{{.Name",
-		},
-		{
-			name:        "valid empty template",
-			expectError: false,
-		},
-		{
-			name:         "valid template",
-			expectError:  false,
-			fqdnTemplate: "{{.Name}}-{{.Namespace}}.ext-dns.test.com",
-		},
-		{
-			name:         "valid template",
-			expectError:  false,
-			fqdnTemplate: "{{.Name}}-{{.Namespace}}.ext-dns.test.com, {{.Name}}-{{.Namespace}}.ext-dna.test.com",
-		},
-		{
-			name:                     "valid template",
-			expectError:              false,
-			fqdnTemplate:             "{{.Name}}-{{.Namespace}}.ext-dns.test.com, {{.Name}}-{{.Namespace}}.ext-dna.test.com",
-			combineFQDNAndAnnotation: true,
-		},
-		{
-			name:             "non-empty annotation filter label",
-			expectError:      false,
-			annotationFilter: "kubernetes.io/ingress.class=nginx",
-		},
-	} {
-		t.Run(tt.name, func(t *testing.T) {
-			_, err := parseTemplate(tt.fqdnTemplate)
-			if tt.expectError {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestFqdnTemplate(t *testing.T) {
-	tests := []struct {
-		name          string
-		fqdnTemplate  string
-		expectedError bool
-	}{
-		{
-			name:          "empty template",
-			fqdnTemplate:  "",
-			expectedError: false,
-		},
-		{
-			name:          "valid template",
-			fqdnTemplate:  "{{ .Name }}.example.com",
-			expectedError: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			tmpl, err := parseTemplate(tt.fqdnTemplate)
-			if tt.expectedError {
-				assert.Error(t, err)
-				assert.Nil(t, tmpl)
-			} else {
-				assert.NoError(t, err)
-				if tt.fqdnTemplate == "" {
-					assert.Nil(t, tmpl)
-				} else {
-					assert.NotNil(t, tmpl)
-				}
-			}
-		})
-	}
-}
-
 type mockInformerFactory struct {
 	syncResults map[reflect.Type]bool
 }
